organization and some custom shader work
