+++ Assets/Materials/Pixelated.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c9492740464bc34d973d3699c10561f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
