+++ Assets/Materials/Test Shader Mat.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d154a48f2d2bd7d4985b349ebea38a98
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
