+++ Assets/Scripts/Enemy/Bullet.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8e46767c5d58a824db399d2d1de9a7cf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
