+++ Assets/Scripts/Enemy/EnemyPath.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 049ab50b363ee5142836a3615641f49a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
