+++ Assets/Scripts/LevelSelector/CustomSplineAnimate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7d67112bd33070d45866fcae83bbb852
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
