+++ Assets/Scripts/Managers/PlanetManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d5c2ba89feba12b4ca15eb92b03fb2d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
