+++ Assets/Scripts/Planet.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 823b48eba3b24a74bab712c9254a8a0c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
