+++ Assets/Scripts/Planet/PlanetEventReceiver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e190d106402daef4d9331348d7d68522
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
