+++ Assets/Scripts/Sun/SunDamage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7b17f7d95fc2648439b3b98a823baa9c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
