+++ Assets/Scripts/Sun/SunDamageVisual.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d608ebdfa04204f4c9ba1d7468de4fdb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
