+++ Assets/Shaders/Pixelated.shadergraph.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: aca05e698f4aeac4282734702c0dcf95
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 11500000, guid: 625f186215c104763be7675aa2d941aa, type: 3}
