+++ Assets/Shaders/Planet.shadergraph.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 971c2c43bb586dd42ac42ceb27e0d9a6
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 11500000, guid: 625f186215c104763be7675aa2d941aa, type: 3}
