+++ Assets/Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e4dbb40e5815952499db406b0f77074e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
